pyproject.toml
[project]
name = "image2characters"
version = "0.0.1"
description = "Transfer image into characters."
authors = [
    {name = "TaoChenyue", email = "chenyue.Tao@qq.com"},
]
dependencies = [
    "pillow>=10.3.0",
    "numpy>=2.0.0",
    "opencv-python>=4.10.0.84",
    "tqdm>=4.66.4",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.2",
]
